refactor/#25/letter 도메인의 dto폴더 이름 dtos로 수정
