+++ src/letter/letter.swagger.ts
@@ -1,9 +1,9 @@
 import { applyDecorators } from '@nestjs/common';
 import { ApiBody, ApiOperation, ApiResponse } from '@nestjs/swagger';
-import { SendLetterDto } from './dto/send-letter.dto';
-import { SaveWritingDto } from './dto/save-writing.dto';
-import { ResLetterDto } from './dto/res-letter.dto';
-import { DeleteLettersDto } from './dto/delete.letter.dto';
+import { SendLetterDto } from './dtos/send-letter.dto';
+import { SaveWritingDto } from './dtos/save-writing.dto';
+import { ResLetterDto } from './dtos/res-letter.dto';
+import { DeleteLettersDto } from './dtos/delete.letter.dto';
 
 export const ApiLetters = {
   send: () =>
